# openapi/glenkeos-complete-api-v1.yaml
openapi: 3.0.3
info:
  title: GlenKeos Platform API
  version: 1.0.0
  description: |
    Complete API specification for the GlenKeos multi-tenant commerce platform.
    Includes all 10 microservices, PayPal integration, and event-driven workflows.
  contact:
    name: GlenKeos Engineering
    email: engineering@glenkeos.com
servers:
  - url: https://api.glenkeos.com/v1
    description: Production
  - url: https://staging-api.glenkeos.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local Development
security:
  - CognitoAuthorizer: []
components:
  securitySchemes:
    CognitoAuthorizer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
        store_id:
          type: string
        status:
          type: string
          enum: [PENDING, ACCEPTED, IN_PREP, READY, OUT_FOR_DELIVERY, COMPLETED, CANCELLED]
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        total_amount:
          type: number
        payment_status:
          type: string
        created_at:
          type: string
          format: date-time
    OrderItem:
      type: object
      properties:
        id:
          type: string
        item_id:
          type: string
        quantity:
          type: integer
        price:
          type: number
        customizations:
          type: object

    # Payments
...
paths:
...
  /orders:
    post:
      summary: Create order
      operationId: createOrder
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customer_id, store_id, items]
              properties:
                customer_id:
                  type: string
                store_id:
                  type: string
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      item_id:
                        type: string
                      quantity:
                        type: integer
                      customizations:
                        type: object
                delivery_address:
                  type: object
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

    get:
      summary: List orders
      operationId: listOrders
      tags: [Orders]
      parameters:
        - name: customer_id
          in: query
          schema:
            type: string
        - name: store_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
  /orders/{orderId}:
    get:
      summary: Get order by ID
      operationId: getOrder
      tags: [Orders]
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /orders/{orderId}/status:
    put:
      summary: Update order status
      operationId: updateOrderStatus
      tags: [Orders]
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_status]
              properties:
                new_status:
                  type: string
                  enum: [ACCEPTED, IN_PREP, READY, OUT_FOR_DELIVERY, COMPLETED, CANCELLED]
                reason_code:
                  type: string
      responses:
        '200':
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

  # ========================================
  # PAYMENTS SERVICE
  # ========================================
...
  /payments/orders/{orderId}:
    get:
      summary: Get PayPal order details
      operationId: getPayPalOrder
      tags: [Payments]
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PayPal order details
          content:
            application/json:
              schema:
                type: object
...